refactor(dashboard): a rail to one destination is no rail, so the header carries the account
The only nav item was the page it linked to. NavUser was the sole way to sign
out, so it moves into the header as an avatar menu, and the Apps list grows the
heading the header used to render for it.
