--- apps/dashboard/src/components/app-sidebar/nav-user.tsx
@@ -1,65 +1,0 @@
-import { EllipsisVerticalIcon, LogOutIcon } from 'lucide-react';
-import { UserIdentity } from '#components/app-sidebar/user-identity.tsx';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '#components/ui/dropdown-menu.tsx';
-import {
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-  useSidebar,
-} from '#components/ui/sidebar.tsx';
-import { useSession } from '#lib/hooks/use-session.ts';
-import { useSignOut } from '#lib/hooks/use-sign-out.ts';
-
-export function NavUser() {
-  const session = useSession();
-  const signOut = useSignOut();
-  const { isMobile } = useSidebar();
-
-  if (!session) {
-    return null;
-  }
-
-  const { user } = session;
-
-  return (
-    <SidebarMenu>
-      <SidebarMenuItem>
-        <DropdownMenu>
-          <DropdownMenuTrigger
-            render={<SidebarMenuButton size="lg" className="aria-expanded:bg-muted" />}
-          >
-            <UserIdentity user={user} />
-            <EllipsisVerticalIcon className="ml-auto size-4" />
-          </DropdownMenuTrigger>
-          <DropdownMenuContent
-            className="min-w-56"
-            side={isMobile ? 'bottom' : 'right'}
-            align="end"
-            sideOffset={4}
-          >
-            <DropdownMenuGroup>
-              <DropdownMenuLabel className="p-0 font-normal">
-                <div className="flex items-center gap-2 px-1 py-1.5 text-left text-sm">
-                  <UserIdentity user={user} />
-                </div>
-              </DropdownMenuLabel>
-            </DropdownMenuGroup>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem onClick={() => signOut.mutate()} disabled={signOut.isPending}>
-              <LogOutIcon />
-              Sign out
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </SidebarMenuItem>
-    </SidebarMenu>
-  );
-}
