--- apps/dashboard/src/lib/hooks/use-mobile.ts
@@ -1,22 +1,0 @@
-import { useEffect, useState } from 'react';
-
-// Generated by `shadcn add sidebar`, then rewritten to the hook conventions in
-// AGENTS.md. components.json aims the CLI's `hooks` alias at this folder, so a
-// regeneration lands here — and overwrites these edits.
-
-// Tailwind's `md` breakpoint, which is where the sidebar switches to a drawer.
-const MOBILE_MEDIA_QUERY = '(width < 48rem)';
-
-export function useIsMobile(): boolean {
-  const [isMobile, setIsMobile] = useState(() => window.matchMedia(MOBILE_MEDIA_QUERY).matches);
-
-  useEffect(() => {
-    const query = window.matchMedia(MOBILE_MEDIA_QUERY);
-    const syncFromQuery = () => setIsMobile(query.matches);
-
-    query.addEventListener('change', syncFromQuery);
-    return () => query.removeEventListener('change', syncFromQuery);
-  }, []);
-
-  return isMobile;
-}
