+++ apps/dashboard/src/routes/(dashboard)/apps/index.tsx
@@ -7,7 +7,8 @@ export const Route = createFileRoute('/(dashboard)/apps/')({ component: RouteCom
 function RouteComponent() {
   return (
     <div className="@container/main flex flex-1 flex-col gap-4 p-4 md:gap-6 md:p-6">
-      <div className="flex justify-end">
+      <div className="flex flex-wrap items-center justify-between gap-3">
+        <h1 className="font-medium text-base">Apps</h1>
         <DeployDialog />
       </div>
       <AppsList />
